majorUpgrades.ContentUpgrades.MinorTweaks/bugfixes 2.74.285 added decimal support to mining added cauldrons to toolsmith trades changed milk quantity from 1->5 in farmer trades optimized drop logic to culinary cauldron and culinary delights cooking logic, allowing for the items to drop more naturally and in the cauldrons case, even launch towards the player that cooked. fixed a bug where non hired villagers could do workcycles

+++ src/main/java/goat/minecraft/minecraftnew/subsystems/mining/Mining.java
@@ -254,12 +254,12 @@ public class Mining implements Listener {
 
             // Apply haste effect based on Mining level
             int miningLevel = xpManager.getPlayerLevel(player, "Mining");
-            int doubleDropChance = miningLevel / 2;
+            double doubleDropChance = (double) miningLevel / 2;
 
             boolean hasDiamondGem = gemManager.getGemsFromItem(tool).contains(MiningGemManager.MiningGem.DIAMOND_GEM);
-            int tripleDropChance = hasDiamondGem ? 10 : 0; // 10% chance for triple drops if Diamond Gem is applied
+            double tripleDropChance = hasDiamondGem ? 10 : 0; // 10% chance for triple drops if Diamond Gem is applied
 
-            int roll = random.nextInt(100) + 1;
+            double roll = random.nextInt(100) + 1;
 
             if (roll <= tripleDropChance) {
                 // Triple drop chance
